requirements.txt
pyTelegramBotAPI
instagrapi
requests
python-dotenv
